Cleanup service files for packs
